Require slimmer/test to provide us with a jQuery.
This test appears to have broken at af7aa963386ee3eb91ccd06b9c445514d000797f.

Which commits is this test involved in?
$ COMMITS=`git log --pretty=oneline test/integration/loading_places_test.rb \
  | awk '{print $1}'`

Which commits did the test pass/fail in?
$ for COMMIT in $COMMITS; do
  git checkout $COMMIT >/dev/null 2>/dev/null
  echo "Commit $COMMIT"
  bundle exec ruby -I test test/integration/loading_places_test.rb \
    >/dev/null 2>/dev/null
  echo "Status: $? (0 = success)"
  git reset --hard HEAD >/dev/null 2>/dev/null
  git checkout master >/dev/null 2>/dev/null
done

...
Commit af7aa963386ee3eb91ccd06b9c445514d000797f
Status: 1 (0 = success)
Commit b97a424b44b45f2e3f2184f71dd11a53d38cfa2f
Status: 0 (0 = success)
...

(I realised after this that I should have just used `git bisect` to find the
commit).

There are still a lot of Javascript errors in the tests, but none of them
cause a test failure. Looks like we're not loading some other Javascript that
the page under test needs to work, probably because the Slimmer template used
in the tests doesn't know about it.