Fix issue with primary footer nav not scrolling to top.
There was a race condition where it would look for the tab it was
referencing before scrolling - but the tab wouldn't switch until after
the jQuery live event fired. This checks to see if the footer link only
links to a fragment identifier and if so will jump to the top.