Add basic tests for searching
These are copied from whitehall and then slimmed down. We don't have the document class
in frontend that we used in rummager so that involves some view changes and extra helpers.
We may want to revisit this once we see the effect of this work more broadly.

This branch doesn't currently display specialist results properly. That requires a bit
more exploration.