Add new SearchResult class and use it to consolidate views
There weren't really many differences between the specialist guidance and mainstream
search results. Those differences can begin to be masked by adding a wrapper around
the raw results.

There's more that could be done to tidy this up but it doesn't make sense to do this
when the display of these results is likely to change quite a bit over the coming weeks.