Switch to content api for publication content
Uses a presenter class to wrap the response from the content api to replace the
existing response that comes from the publisher api. This mean very few changes
are needed to the controller and none are needed on the views.

Tests are still failing that relate to local transactions as they are still
being switched over to the content api and are fairly mixed up with this work.