Add campaign pages (initially workplace pensions)
This is the beginning of the campaign pages work. It'll be implemented
as static templates implemented by developers for the time being.

To start with, this is the DWP Workplace Pensions campaign page. The
assets and relevant template have been set up.

At the moment, the route is hard-coded to point through to the
controller method/template.