Use the providing tier for licences where available
If the local service is available for this licence, match the
providing tier to the councils in the geostack to return the most
appropriate authority for the user's location.

If there is not a local service available, or if the user's location
does not include an authority matching the providing tier, fall back
to the existing behaviour of selecting the closest authority to the
user.