Update the search form action hash based on tab.
When you are looking at a tab on search and then change your search
terms and re-submit the form you arrive back on a different tab. This
has been causing lots of confusion for people searching for things on
Inside Government as its not the default tab.

This should make sure the form reflects the tab that you currently
looking at when chaning your search terms.