Build an initial part if no parts are present
For travel advice, the content API will return a
successful response with country details regardless
of whether a published TravelAdviceEdition exists
or not.

In this commit, we catch when this is the case
(where @publication.parts is not present) and build
a stub "summary" part so that we can still display
a skeleton travel advice page for the country.