Split apart guide navigation templates
Despite the additional logic added to the guide
navigation templates, there is now a need to add 
further changes to the travel advice guide navigation.

In light of this, I'm splitting the templates apart
into separate guide/benefit and travel-advice partials 
so we can keep them as simple as possible.