Bump API adapters to 4.3.0.
This gives us the ability to follow redirect responses from our APIs,
which in turn means we can change the APIs' URL structures without
requiring a new release of every dependent app.

The app we're looking to migrate, in this case, is the content API, but
the new feature applies to all APIs.