Tidy up next/previous part logic
Increase the encapsulation by bringing more of this logic into the PublicationPresenter
and thereby make the view code more readable. Also pull some shared view code out into
a partial.

The PublicationPresenter is now becoming quite a complex class, but it remains a more
natural home for this code. We do still need PartHelper but are hopefully reducing our
reliance on it.