Break appropriate_snac_code_from_location into two methods
This method was doing two things: acting as a factory for an identifier
class, and then using it to find the snac. Breaking it into two methods
makes that a little clearer at the expense of increasing the number of
methods in this controller related to snac codes - this may be something
we should extract?