Don't list multiple locations in search results.
Certain publications are associated with a large number of world locations,
and so show up with a great long list in the search results. This is quite
confusing, so we just say "multiple locations" and let people follow through
to the publication if they want the whole list.

This was the initial behaviour, but was changed (possibly inadvertently) in
commit e0a7a6150076ad9de0b9cb59e568535fff7dc629.