Fix previewing of simple_smart_answers
Capybara with rack-test behaves differently to real browsers when
handling method=get forms with extra get params in the action URL.
Real browsers remove the get params, and replace them with the form
values. Capybara (or rack-test) merges the params together.  This was
causing this test to give a false positive.

PhantomJS (via Poltergeist) behaves in the same way as a real browser.