Gracefully handle when a sector tag does not exist
When a request for an industry sector tag returns a 404 from the Content 
API, return a 404 status so that we render an error page.

I've extracted the 'api_returns_404_for' method from the browse 
functional tests into a file in the test/support/ directory so that 
other tests are able to use it.