Make section-wrap use display:table-row
In order to get subscriptions and country filter to vertically align I was setting a height and absolutely positioning the subscriptions to the bottom. While this worked in all browsers it was really yucky so I've rewritten to use display:table-cell and vertical-align. This solution has fewer gotchas for future contributors to this file (I hope).

In ie-lte8 the subscriptions sits under the search.