Set a cookie based on the position of link clicked
So that we can find out where users are clicking on search results we
are going to set a cookie containing the position of the link. We will
then save the position information to Analytics on the next page.

This is done via a cookie and on the future page so we don't block
leaving the page while we call off to analytics.