Refactor specialist sector routing
This changes how routing works for specialist sectors, so that we only 
have to define them in a single place. A definitive list of specialist sector tags has been added in lib/frontend.rb. 

The routes configuration has been updated to use this list when setting 
constraints on the specialist sector routes, and the Rake task 
"router:register" registers prefix routes for each sector in the list.