Content Api raises an exception with Artefact errors
With the update to gds-api-adapters in this branch, we saw some errors
where Frontend expected to see a 410 if the Artefact had been archived.
This update to artefact_retriever uses the new .artefact! method
introduced with
https://github.com/alphagov/gds-api-adapters/commit/90ee3904 which will
raise an exception for both 404 and 410.
This commit also refactors all the tests that stubbed out .artefact or
had issues with the exceptions being raised.