Explain why we use a FORM here
Specifically, we use a FORM because of Dragon Dictation - users are
likely to say something like "click the start button", which won't work
if it's a link as Dragon doesn't obey ARIA directives right now.

Also, we use a GET because the downstream service subdomain 405s POSTs
to its start page (as it should, really).