Stop redirecting /search.json to the Content API
The new unified search controller has a JSON representation which is 
used to load results in-place on the page. When we remove the 'ui' 
feature-flag, the JSON route for unified search will conflict with 
the existing redirect route.

There are almost zero requests to /search.json in its current form, and 
we recently changed the structure of the API response at the redirect's 
destination without any complaint from users, so I don't think changing 
this route is going to pose a problem.