Make the unified search the default
This changes routes for search so that the unified search action is 
now the default, and requests for search will only hit the previous 
tabbed interface when the 'ui' parameter is 'old'.

This updates the functional tests for the UnifiedSearchController to 
no longer provide the 'ui' parameter, and also prevents the 'ui' param
from leaking out into form requests on subsequent searches.