Send external search results through link tracker
So that we can see if people are clicking external links and if they are
producing any value.

Implemented in the same way that the link tracking/redirection works on
Google by switching the href of the link on mousedown or keydown. This
means that all normal browser will be intact like opening in a new
tab/window.

Ideally we will use the beacon api in the future but it isn't
very well supported yet:
https://developer.mozilla.org/en-US/docs/Web/API/navigator.sendBeacon