Call imminence directly for find_nearest searches.
This change takes content_api out of the path for finding nearest places from a
postcode, as part of preparations for the move to the new content store.

Previously frontend was calling MapIt to get lat/long for a postcode, then
passing that location to content_api which in turn called imminence to get the
nearest places. Imminence now has the ability to call MapIt, so this change
removes the MapIt call here and instead calls imminence with the postcode,
receiving a list of places.