Pass the debug parameter through to rummager
This parameter allows the ranking options to be changed; for example, it
allows use of best-bets to be turned off.  It's useful to be able to
make these modifications on the live search results to compare the final
ranking that would be produced, and to check if options like popularity
information help for particular searches.