Mark closed organisations in search results
Requires requesting the list of fields to be returned from search, since
the default list doesn't include the organisation_state field.  This is
something I've wanted to start doing for a while, anyway (I'd like to
remove the default entirely, so that apps only request what they
actually want to use).

Content designer feedback suggested that the appropriate place to mark
the org as closed is at the start of the title.  We also don't want to
add "The home of <organisation> on GOV.UK." for closed organisations,
since it's misleading (the organisation arguably doesn't have a home any
more).

Related story: https://www.pivotaltracker.com/story/show/75035036