Make browse links alphabetical on homepage
So that they match the order of the new browse pages and give people a
little more context as to what they are looking at when they navigate
through.

Done as three lists which is less than ideal but due to the limitations
of CSS floats I can't think of any better ways to achieve this.