Add a visually hidden result count
For aria-live to work, the bit that is read out needs to remain on the page (not be swapped out by the javascript). Putting a visually hidden result count that remains on the page and is then updated by the javascript means that when the result set has finished updating screen readers will read out the new result count.