SearchResultsPresenter returns prev and next links
This changes the SearchResultsPresenter class to calculate and return
the pagination links for search results. 

The class exposes methods to return the next and previous page link, 
labels for the links which provide page numbers for context, and two 
boolean methods to assert whether there is a next or previous page 
present.

These new methods are exposed in the to_hash method, so that they can 
be easily used in the Mustache templates.