Force the start parameter to be a string
It appears that, when the 'start' parameter is a Fixnum, it does not 
get included in the params of the Rummager request. For now, explicitly 
converting this to a String fixes the issue. 

In the future, we might want to look at fixing this in API adapters, 
however given the work to fix the cruft in URL building (cc @elliotcm) 
it's probably not worth looking into right now.