Add cross-domain analytics for Civil Claims start page
The cross domain analytics behaviour was first added for IER (see
https://github.com/alphagov/frontend/pull/643) for original
discussion, this expands it to also work for Civil Claims.

This is a simplest-thing-that-works approach, pushed a little bit
further.

Manually targeting the couple of services in the template, as we
don't know enough to generalise yet, nor is it worth the effort
of making this part of publisher - though probably the next
time we get a request we should consider it (weighing up against
the planned move for GOV.UK to GA Universal, which may affect
this implementation)