Temporarily remove DVLA phone contact on view-driving-record
DVLA are getting a high level of calls from people calling the number
on this transaction when they calls aren't relevant, and can't currently
be re-routed.

Until then remove the contact number, as this number can't help answer
(general) questions that are being asked.

Once DVLA can re-route this will be reverted.