Remove the tax disc A/B test
The tax disc A/B test has been used to 
incrementally increase the number of visitors who
are presented with the Vehicle Tax beta as a
primary option.

The beta has been running at approx. 100% for the
last 3 months. We can now remove the A/B test and
make the change more permanent.

This commit removes the A/B test from the tax disc
start page and any associated assets.