Remove links to the original tax disc service
The original tax disc service doesn't support the new Direct Debit
payment options and is missing some other more recent changes. The
majority of users have been using the beta service since August 2014
and this is where DVLA's development effort is now focussed.

As the original service was disabled by DVLA on 30 September, we should
remove links to it to avoid confusing users.