Fix double-inclusion of Universal Analytics
Since https://github.com/alphagov/static/pull/508 was merged the
Universal Analytics code used to support cross-transaction tracking
would be broken.

This commit updates the in-template code as follows:

- Remove dynamic creation of Javascript loader
- Use UA namespacing[0] to namespace the `pageview` beacon sending
  functionality

This commit will need to be released at the same time as the code in the
above pull request to avoid any production issues.

[0] https://developers.google.com/analytics/devguides/collection/analyticsjs/advanced#multipletrackers