Display a generic description for topics in search
We'd like to return topics (which used to be called specialist_sectors
and still are in the format name) in search.  However, these have no
metadata or description, so they look quite confusing in search.  To
make them actually look like a search result, this commit adds a generic
description for each topic result which is missing a description.