Fix gutter spacing on campaign pages
At the moment, the spacing in the gutters of campaign pages pull the 
page too far towards the left on desktop. On mobile, the padding is too
much. In both cases, the edges of the campaign page do not line up with 
the GOV.UK logo and search box in the header.

This fixes the padding based on values used for quick answers and 
guides.