Show groups in search results
This is the frontend half of some work to group search results by facets
such as topics, mainstream browse pages, or organisations.  It will have
no visible effect without corresponding changes in rummager, so is safe
to merge and deploy before the rummager changes are made.

After this change, Rummager can suggest groupings of search results in
the search responses.  These are represented by a result document with
an extra member, "examples", containing a list of example items in the
group.

Rummager may also suggest a filter along with a search result, which
will be converted to a link to apply that filter.  This allows users to
see everything in a group matching a search.

I'd like to get this merged so that I can experiment with the support in
rummager (under the control of a feature flag).