Add metadata to non-edition search results
Previously there were two classes of search results: results from the
mainstream index, typically things like guides and smart answers which
have browse sections as metadata, and results from the government index
which have date, organisation and format as metadata.

We're slowly moving towards a model where everything is added to the
mainstream index, and the new existing metadata approach doesn't work
for these new content items. These new results are being added with a
specific `document_type`, instead of having everything lumped together
as an `edition`.

This commit adds a third type of search result presenter which is used
for any results that are in the mainstream index and have a type other
than `edition`. It shows a date, a humanized version of the document
format, and the organisations the result is tagged to.

The humanized version of the document format is set manually for a few
types because they contain acronyms which aren't properly handled by
Rails' `#humanize` method.

Most of these document formats have `last_update` populated instead of
`public_timestamp`. Both fields have the same semantics, so we make the
presenter prefer `public_timestamp` and fall back to `last_update` when
unavailable. Over time we will make all formats use `public_timestamp`
properly and then remove `last_update`.