Enable search filtering by topic
This is a "hidden" feature for now.  If `filter_topic[]=foo` is added to
the search query parameters, a second filter box is shown in the search
results, and the results are filtered to show only those matching the
requested topics.

The internal field name used for these is "specialist_sectors", but we
don't want to expose that name since it's misleading, and we're moving
to call these "topics" everywhere.  The search index will probably be
moved to the new field name when we rebuild it to listen to the
messaging queue.

There are a few rough edges, which I don't think need to block this
being deployed:

 - If filters are selected in one of the filter boxes, the options shown
   for the other field (in the other filter box) should update.
 - If all filters are removed from the "topics' filter box, and then the
   page is refreshed, the topics filter box will disappear.