Preserve active filters with spelling suggestions
When a user clicks on one of our spelling suggestions, we were dropping
any filters which were active on the search.  This is generally
unhelpful, but particularly problematic if the search had filters
automatically added (for example, due to searching from one of the
organisations for which scoped search is automatically enabled).

Instead, we now preserve the filters.

Also, record the original query that was manually entered in the "o"
query parameter.  This matches the parameter used (at least
historically) by other search engines such as Google for this purpose.
This parameter is ignored by frontend, but we can use it in analytics to
track how often the suggestions are being used (and which result is
eventually clicked on when it is used).