Avoid linking to search parameters with start=0
When we generate links to search results, if the calculated start value
is 0 don't add a start parameter (since this is the default anyway).

This makes URLs nicer, but also helps caching.  For example, if someone
paginates to page 2, then the link to page 1 would go to a cached page.

I've checked with Tara, and this won't make any analytics significantly
more awkward.