Don't add metadata to group results
For groups of results returned from Rummager (not yet in production),
the format is "group", but we don't want to show this as part of the
result.  Add a class for GroupResults (it currently has nothing that
SearchResult doesn't have, but it's useful to mark it as different,
since it has a very different look), and use this class instead of the
NonEditionResult for group results.