Fix layout on travel advice index and search pages
The search page needed turning into a full width page as its main
element should be the full width of the page.

Both pages also needed removing of some padding from the side of the
pages as they were accounting for the wrapper to not bring them in to
the site width.