Add a Content-Security-Policy reporting header
Add `Content-Security-Policy-Report-Only` to the homepage in
development only.

Whitelist:

- static and assets hostnames for loading anything
- Google Analytics for loading JavaScript and fallback images
- Inline JavaScript and styles (we have a lot of existing inline JavaScript
  and inline styles loaded with jQuery)
- Data URIs for loading our font files
- Reporting to the event-store app

There is still one report firing for me that I haven't been able to
get rid of:

> Refused to load plugin data from ''

This is related to the accessible video player, which tests Flash
on every pageload in JavaScript. I don't think there's a way to
allow this behaviour.