Bump slimmer and update tests for component label
Bump slimmer so that we get the latest test helpers for the components.
The other change that comes in with the version bump is that report a
problem will now appear on every page regardless of if it sets an empty
report a problem element or not. Because the report a problem is on
every page by default remove all the tests to assert it has been added.