Remove article tags and reliance on them
Use article tags in a more semantic manor. There are two parts to this.
In the first instance I have removed a whole collection of questionable
article tags. The category for questionable was do they contain a
heading. The second was add a `content-block` to any article tags which
I couldn't make an easy decision on. Adding this class will enable us to
keep the styling when we remove the article tag from the CSS selector.