Change /random cache from private to public
By default `expires_in` will generate a `Cache-Control` header with the
`private` instruction, which allows the client to cache the response but not
any intermediate proxies such as our Varnish nodes or CDN:

- http://apidock.com/rails/ActionController/ConditionalGet/expires_in
- https://github.com/alphagov/cdn-acceptance-tests/blob/e4ee66728c99cedb8b510c7736df35b4ee5b6bd4/cdn_nocache_test.go#L71-L81

This means that we would probably see the same amount of traffic at and
within origin, especially from bots. Change the instruction to `public`
which will allow the response to be cached by those layers and soak up any
brief increases in traffic.

I've moved the comment about why we're doing this, minus the specific value,
to an additional test which asserts this behaviour is happening. This should
reduce the risk of the code diverging in behaviour from the comments. The
move of where we call `expires_in` is purely aesthetic.