Hide services and information heading on homepage
This heading isn't particularly illuminating, and makes the mainstream
browse entry points appear further down the page than they need to.  We
think the heading is still useful for screenreaders, so we just mark it
visuallyhidden.

We use the phrase "Services and Information" on (some) organisation
pages to mean something quite different (a link to a set of subtopics
containing documents related to the organisation), so it would be good
to avoid using the same title in both places.  I therefore tried to find
alternative wordings for the heading that will be clearer when read by
screenreaders, but we haven't been able to come up with anything
reasonable.  It's possible that the heading "Services and Information"
is appropriate here, and it's the label on organisation pages that could
do with changing.