Handle campaigns without organisations
Organisation HTML was being included in the page even for campaigns
without an organisation. The conditional that checked whether the
campaign had an organisation was always returning true, because there
was always a hash with empty values for the organisation details.

To fix this we added a `has_organisation?` method which checks if any
of the organisation values are present.

It's still possible to get weird HTML included on the page, if the
organisation details contain some values but not all of them. We've
spoken to the content team and they're happy to ensure
organisationless campaigns have no values set for organisation.