Removing div with inner class from _publication_metadata
When working on something else, as an aside, I noticed this commented
out div which has been commented out since May 2013. My guess is that
we can remove this as it's not being used. I have searched frontend
repo for .inner styles nested inside .meta-data or .group classes
and couldn't find any that required deleting along with the removal
of this commented out code.

This removal will not change anything and does not require any before
and after screenshots