Update live search to use universal and classic page tracking
We are currently in the process of moving over to use universal
analytics. This change updates live search to use both classic and
universal. The plan is to get universal analytics monitoring exactly
what classic currently does.

This change uses GOVUK.analytics from Paul hayes recent work on static
https://github.com/alphagov/static/pull/549
As suggested by Paul we removed the last figure of 3 because that is
the default figure for that and is not needed.

https://www.pivotaltracker.com/story/show/87738926