Add history mode note to search results
When a search result is flagged as `historic` we want to highlight
this to a user browsering search results, so they know that result
may not be up to date and may not reflect the views of the current
government.

`historic` is defined as `political` content published under a
previous government. This only applies to Whitehall Editions
currently.