Wrap rummager api with own client
In order to get information about the scope (specifically, we want the title of the document we're searching in here, but in the future we might want other stuff) create a wrapper around the rummager api so we can make multiple requests to rummager which the controller doesn't need to know about. This decouples our controller from rummager so that when rummager can take on more of this work from the search_api we don't need to change the controller at all.

This commit:
 - Adds a `SearchAPI` around rummager
 - fixes some long-standing broken tests where the test stubs Facets to be a hash not an array (these tests only just started failing since we started using facets)
 - Fix the failing controller tests, which since we changed the controller to use a wrapper, now is not stubbing the appropriate thing. To fix them I've just removed the expectation of what the controller passes in to the api wrapper. Since the controller doesn't do anything with `search_params` object except instantiate it, testing that it is the right thing is not very useful. Tests like this should go into unit tests on the `search_parameters` class instead. (This is a piece of work for a later commit)
 - Add unit tests for newly created SearchAPI. These tests
 -- Tests that scoped searches return the scope and the search results
 -- Tests that un-scoped searched returns just the results
 - Replace stub method on search results presenter with code to use newly returned manual title