Add request for extra results
When users search within a manual (this is doing a "scoped" search)
we want to show them some results from the rest of GOV.UK (ie some
"un-scoped" results) so that they are aware that here are other results
available that might be useful.

This commit adds a third rummager request for scoped searches which
returns three un-scoped results. This request is built by removing
the scoping `filer_manual[]=this/manual` and adding an extra parameter
`reject_manual[]=this/manual` which tells rummager we don't want any
results that occur in this manual.