Use 'invalid_utf8_rejector' to remove malformed UTF-8 paths
The 'invalid_utf8_rejector' inserts rack middleware in front
of the app which returns 400 to requests with malformed UTF-8 paths
(not this is slightly different to the 404 status that the app
returns before this commit). The tests are being removed because
malformed requests should no longer make it to the app.

This gem is already in use within Whitehall.