Change the next and previous buttons to component
This is done for guides and programmes. This change involves
moving the next and previous buttons out of the content-block
and into the footer.

What this changes:

- The "title" attribute on the next and previous links has been removed.
The link title duplicated the link text and wasn't any more helpful
for screen readers.
- The link text ("Next", "Previous") is localised when appropriate, instead
of the link title attribute.
- The "Last Updated" timestamp is now above the navigation buttons and the
print link. This has been reviewed by @wryobservations. The position of the
"Last Updated" text will be revisited when page history for mainstream content
is reviewed.