Display 20 results on search page
Since the introduction of pagination there is no need to show an
excessive amount of results on the search page.

This commit changes the maximum number of results to 20.

Some tests had to be updated to reflect this, because they either
relied on the default value (they don't anymore) or relied on the fact
that when searching with the default number of results we don't show
the `count` in the URL.

From now on, changing the default number of results to anything but 50
wil *not* make the tests fail.