Isolate search parameter tests
Some tests in search_results_presenter_test.rb are currently
duplicated. They are supposed to verify the bounds on `count` and
`start` are being enforced by `SearchParameter`.

However, after this the `SearchResultsPresenter` is tested with the
changed params. This is unnecessary because there are other tests with
the exact same input.

Example: `SearchParameters.new(count: 50, start: -1) ==
SearchParameters.new(count: 50, start: 0)`, and is tested on
search_results_presenter_test.rb#105.