Remove tracking of search results and suggestions
This code isn't using the correct naming, and wasn't tracking the search
results shown either, so we're disabling it so other frontend changes
can be deployed.  It is going to come back in a fixed form shortly, so
I'm not removing the implementation of `extractSearchURLs`, which seems
to be working fine (and has tests), but won't be being run after this is
merged.