Extract out how the search event object is built
Separate out the building of the search result object into a separate
function so that it can be tested seperately from the sending of the
event.

Previously, the event was sending a JSON obect which contained an Array
of URLs that were JSON.stringified jQuery
objects. `buildSearchResultsData` converts these into a simpler Array so
that the URLs are easier to parse.