Render a custom logo for Environment Agency
The Environment Agency have an exemption from the single identity brand
and as such have their own logo which they want to use on their campaign
page. This adds a way to have custom image logos for departments.

This would ideally be using a component to render all logo images so
that we wouldn't need to keep two copies of department logos around the
place. However, that has been deemed to be out of scope for this bit of
work. This is the smallest amount of work needed to get the logo working
right now and we can look at department logos as part of a bigger bit of
work in the future.