Don't treat empty strings as valid analytics profiles
Publisher will happily store a blank string for the department_analytics_profile,
which is then passed to the frontend. Because we were only checking if the value
was truthy, and because an empty string is truthy, we were rendering the
cross domain tracking code when we shouldn't.

Any edition edited in Publisher after we added the field has a blank string for
the department_analytics_profile.

Neither our Javascript nor the Google Analytics client code checks that the
value isn't empty, so it does its thing and appends a Google Analytics parameter
to the link the user follows. This is bad in of itself.

In the case of https://www.gov.uk/apply-online-for-student-finance, we link to
a URL which they redirect, but the extra parameter broke the redirect on their
server and so users saw an error page instead of reaching the destination.