Fix simple smartanswers layout
Issue: simple smartanswers questions are not aligned with the
prominent in mobile view.

This is because section.content does not have any styles applied.
Change to main.content (that is used all around the application) that adds
correct padding. Remove .current css class from .step as it doubles the padding.