Remove old transactions from new window list
This list was created in 2012 to force some legacy transactions to open
in new windows when the “start now” button was clicked. This was because
some of those transactions had either been designed to be opened in new
windows or broke when opened with toolbars enabled.

This commit cleans up the list of transactions by removing those which
no longer exist in their original form. These transactions have either
been rebuilt or closed down.